Module Bitvec_order.Comparators

Open this module to make the following fields available

type bitvec_order = natural

the default ordering for bitvectors

val bitvec_compare : t -> t -> int

bitvec_compare x y orders x and y in the natural order

val bitvec_equal : t -> t -> bool

bitvec_equal x y is true if x is equal to y.

val bitvec_order : ( t, natural ) comparator

natural the packed comparator that sorts in the natural order

val bitvec_ascending : ( t, ascending ) comparator

ascending the packed comparator that sorts in the increasing order

val bitvec_descending : ( t, descending ) comparator

descending the packed comparator that sorts in the decreasing order